+++ changelog.md
@@ -8,6 +8,7 @@
 
 <https://1drv.ms/f/s!ArwKS_ZUR01gySQIE1QCL2i9F_uR>
 
+- icons added and credits in about dialog updated, thanks to: Freepik, vanontom, nulledone, ilko-k
 - fix: if in the codec dialogs search field the enter key is pressed in order to cycle
   there is no longer an annoying sound made 
 - fix: encoding was failing when a ambersand (&) character was used in the path